feat: upgrade to cpp client 3.8.0

+++ Sources/Pulsar/Result.swift
@@ -169,6 +169,6 @@ extension Result {
 	}
 
 	var cxxValue: _Pulsar.Result {
-		_Pulsar.Result(rawValue: Int32(self.rawValue))
+		_Pulsar.Result(rawValue: Int8(self.rawValue))
 	}
 }
